// src/entities/pericias/model/PericiaNome.ts
export enum PericiaNome {
  ACROBACIA = "Acrobacia",
  ADESTRAMENTO = "Adestramento",
  ATLETISMO = "Atletismo",
  ATUACAO = "Atuação",
  CAVALGAR = "Cavalgar",
  CONHECIMENTO = "Conhecimento",
  CURA = "Cura",
  DIPLOMACIA = "Diplomacia",
  ENGANACAO = "Enganação",
  FORTITUDE = "Fortitude",
  FURTIVIDADE = "Furtividade",
  GUERRA = "Guerra",
  INICIATIVA = "Iniciativa",
  INTIMIDACAO = "Intimidação",
  INTUICAO = "Intuição",
  INVESTIGACAO = "Investigação",
  JOGATINA = "Jogatina",
  LADINAGEM = "Ladinagem",
  LUTA = "Luta",
  MISTICISMO = "Misticismo",
  NOBREZA = "Nobreza",
  OFICIO = "Ofício",
  PERCEPCAO = "Percepção",
  PILOTAGEM = "Pilotagem",
  PONTARIA = "Pontaria",
  REFLEXOS = "Reflexos",
  RELIGIAO = "Religião",
  SOBREVIVENCIA = "Sobrevivência",
  VONTADE = "Vontade",
}