// src/entities/atividades/model/AtividadeNome.ts
export enum AtividadeNome {
  EQUILIBRIO = "Equilibrio",
  ESCAPAR = "Escapar",
  PASSAR_POR_INIMIGO = "Passar Por Inimigo",
  CORRIDA = "Corrida",
  ESCALAR = "Escalar",
  NATACAO = "Natação",
  SALTAR = "Saltar",
  CONDUZIR = "Conduzir",
  GALOPAR = "Galopar",
  MONTAR_RAPIDAMENTE = "Montar Rapidamente",
  PRIMEIROS_SOCORROS = "Primeiros Socorros",
  BARGANHA = "Barganha",
  MUDAR_ATITUDE = "Mudar Atitude",
  PERSUACAO = "Persuasão",
  DISFARCE = "Disfarce",
  FALSIFICACAO = "Falsificação",
  FINTAR = "Fintar",
  INSINUACAO = "Insinuação",
  INTRIGA = "Intriga",
  MENTIR = "Mentir",
  ESCONDER_SE = "Esconder-se",
  SEGUIR = "Seguir",
  ASSUSTAR = "Assustar",
  COAGIR = "Coagir",
  PERCEBER_MENTIRA = "Perceber Mentira",
  INTERROGAR = "Interrogar",
  PROCURAR = "Procurar",
  OBSERVAR = "Observar",
  OUVIR = "Ouvir",
  ACAMPAMENTO = "Acampamento",
  IDENTIFICAR_CRIATURA = "Identificar Criatura (CD 15 + ND da Criatura)",
  ORIENTAR_SE = "Orientar-se",

  AMORTECER_QUEDA = "Amortecer Queda",
  PASSAR_ESPACO_APERTADO = "Passar por Espaço Apertado",
  ACALMAR_ANIMAL = "Acalmar Animal",
  MANEJAR_ANIMAL = "Manejar Animal",
  APRESENTACAO = "Apresentação",
  IMPRESSIONAR = "Impressionar",
  IDIOMAS = "Idiomas",
  INFORMACAO = "Informação",
  CUIDADOS_PROLONGADOS = "Cuidados Prolongados",
  NECROPSIA = "Necropsia",
  TRATAMENTO = "Tratamento",
  ANALISAR_TERRENO = "Analisar Terreno",
  PLANO_ACAO = "Plano de Ação",
  PRESSENTIMENTO = "Pressentimento",
  APOSTAR = "Apostar",
  ABRIR_FECHADURA = "Abrir Fechadura",
  OCULTAR = "Ocultar",
  PUNGA = "Punga",
  SABOTAR = "Sabotar",
  DETECTAR_MAGIA = "Detectar Magia",
  IDENTIFICAR_ITEM_MAGICO = "Identificar Item Mágico",
  IDENTIFICAR_MAGIA = "Identificar Magia (CD 15 + Custo em PM da Magia)",
  LANCAR_MAGIA_ARMADURA = "Lançar Magia de Armadura (CD 20 + Custo em PM da Magia)",
  ETIQUETA = "Etiqueta",
  CONSERTAR = "Consertar",
  FABRICAR = "Fabricar",
  IDENTIFICAR = "Identificar",
  SUSTENTO = "Sustento",
  RITO = "Rito",
  RASTREAR = "Rastrear",
}